% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osname.R
\name{os_name}
\alias{os_name}
\title{Human Readable Operating System Name}
\usage{
os_name()
}
\value{
A length one character vector.
}
\description{
The current version was adapted from \code{utils::sessionInfo()}.
}
\examples{
os_name()
}
