% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ osname }
\alias{ osname }
\alias{ osname }
\title{ Human Readable Operating System Names }
\description{
For example, Windows 8.1 instead of version 6.3.9600. On macOS it includes
  the release names, on Linux it includes the name of the distribution.
}
